+ Updated to forge 1.12.2-14.23.1.2596 + Made spawner checking much more efficient + Added a config to filter spawning blocks

+++ src/main/java/com/buuz135/despawningspawners/DespawningSpawners.java
@@ -1,6 +1,7 @@
 package com.buuz135.despawningspawners;
 
 import net.minecraft.init.Blocks;
+import net.minecraft.tileentity.MobSpawnerBaseLogic;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -11,33 +12,28 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
+import java.util.Arrays;
 import java.util.HashMap;
 
 @Mod(modid = "despawningspawners", name = "Despawning Spawners", version = "1.0")
 public class DespawningSpawners {
 
-    private static AxisAlignedBB box = new AxisAlignedBB(-5, -5, -5, 5, 5, 5);
-    private static int MAX_SPAWNS;
-
     @Mod.EventHandler
     public void onPreInit(FMLPreInitializationEvent event) {
         MinecraftForge.EVENT_BUS.register(this);
-        Configuration config = new Configuration(event.getSuggestedConfigurationFile());
-        MAX_SPAWNS = config.getInt("maxSpawnerSpawns", Configuration.CATEGORY_GENERAL, 50, 0, Integer.MAX_VALUE, "Amount of spawns can a spawner do. It won't stop the remaining spawns of a spawner 'work'.");
-        config.save();
     }
 
     @SubscribeEvent
     public void onEntitySpawn(LivingSpawnEvent.SpecialSpawn event) {
-        BlockPos pos = getSpawnerinAABB(event.getWorld(), box.offset(event.getEntity().getPosition()));
-        if (pos != null) {
+        BlockPos pos = getSpawnerinAABB(event);
+        if (pos != null && Arrays.asList(DespawningConfig.spawnerBlocks).contains(event.getWorld().getBlockState(pos).getBlock().getRegistryName().toString())) {
             SpawnerSavedData data = (SpawnerSavedData) event.getWorld().getPerWorldStorage().getOrLoadData(SpawnerSavedData.class, SpawnerSavedData.NAME);
             if (data == null) {
                 data = new SpawnerSavedData();
                 event.getWorld().getPerWorldStorage().setData(SpawnerSavedData.NAME, data);
             }
             HashMap<BlockPos, Integer> spawners = data.getSpawners();
-            spawners.put(pos, spawners.getOrDefault(pos, MAX_SPAWNS) - 1);
+            spawners.put(pos, spawners.getOrDefault(pos, DespawningConfig.maxSpawnerSpawns) - 1);
             if (spawners.get(pos) < 0) {
                 event.getWorld().destroyBlock(pos, true);
                 spawners.remove(pos);
@@ -47,14 +43,9 @@ public class DespawningSpawners {
         }
     }
 
-    public BlockPos getSpawnerinAABB(World world, AxisAlignedBB axisAlignedBB) {
-        for (double x = axisAlignedBB.minX; x < axisAlignedBB.maxX; ++x) {
-            for (double z = axisAlignedBB.minZ; z < axisAlignedBB.maxZ; ++z) {
-                for (double y = axisAlignedBB.minY; y < axisAlignedBB.maxY; ++y) {
-                    if (world.getBlockState(new BlockPos(x, y, z)).getBlock().equals(Blocks.MOB_SPAWNER))
-                        return new BlockPos(x, y, z);
-                }
-            }
+    public BlockPos getSpawnerinAABB(LivingSpawnEvent.SpecialSpawn event) {
+        if (event.getSpawner() != null) {
+            return event.getSpawner().getSpawnerPosition();
         }
         return null;
     }
